requirements.txt
aiohttp>=3.8.0
aiosignal>=1.3.0
attrs>=23.1.0
discord.py>=2.3.0
frozenlist>=1.3.0
google-genai>=0.5.0
idna>=3.4
multidict>=6.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
python-magic>=0.4.27
soundfile>=0.12.0
typing-extensions>=4.7.0
yarl>=1.9.0
yt-dlp>=2025.6.9
watchdog>=4.0.0
